// src/generated/models/userProfileWriteRole.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * One Health Lyme Gap Atlas API
 * Build artifact for existing first-party Web client and validator generation. Not an external developer contract or HTTP surface. Use the public openapi.json for external documentation.
 * OpenAPI spec version: 0.1.0
 */

export type UserProfileWriteRole = typeof UserProfileWriteRole[keyof typeof UserProfileWriteRole] | null;


export const UserProfileWriteRole = {
  general_public_citizen: 'general_public_citizen',
  district_level_epidemiologist: 'district_level_epidemiologist',
  state_level_epidemiologist: 'state_level_epidemiologist',
  state_director_level_epidemiologist: 'state_director_level_epidemiologist',
  national_level_epidemiologist: 'national_level_epidemiologist',
} as const;